Completely revamp the syntax
The previous syntax modeled keywords as `@`-prefixed strings in keys
or values, but this was causing language design problems (e.g. property
name collisions in reasonable programs), and it didn't really leverage
the molecule structure as syntax as originally envisioned.

With the new syntax, keywords must now be `@`-prefixed string values
of properties named `'0'`. Also, arrays are now allowed in input JSON,
so `["@example"]` means the same thing as `{ "0": "@example" }`. The
entire object that the keyword resides within is in scope when the
keyword is applied. Keywords are now always evaluated depth-first.