Accept arbitrary JSON as input
Non-string primitive values are canonicalized to their stringified form
(e.g. `1` becomes `"1"`).

Also, `@`-prefixed strings are now handled similarly whether they are
in keyword position or not. Previously only strings that could possibly
be keywords could use the `@@` escape sequence. Now all strings (both
as keys and values) can. It is now an error to have anything that looks
like a keyword (a string starting with a single `@`) anywhere other
than as the value of a property named `0` (this is reserved syntax
space which may be leveraged in the future).