Improve sanity of trailing newline handling in command-line output
I had confused myself about why a trailing newline was necessary to see
complete command-line output. It's due to an [npm bug][0], not any
behavior of Node itself, and as such ought to be worked around in the
npm scripts.

However, whether this bug is present or not it's still good practice
for most command outputs to end with a complete line (including `\n`),
so all `pretty-*` notations have been updated to include a trailing
newline.

In the future I may decide to add additional `--output-format`s which
use non-`pretty-*` notations so that users can get non-`\n`-terminated
outputs from the runtime, but for now there's nothing new here.

[0]: https://github.com/npm/cli/issues/8583