Allow `SemanticGraph`s to contain `symbol`s as well as `string`s
These `symbol`s are referred to as "qualia" in the implementation as a
dual to "atoms" (`string`s). They cannot be serialized and are meant
for internal usage within the runtime; e.g. to refer to intrinsic
types. There is currently no way for users to create new `symbol`s.